+++ patzilla/access/uspto/pdf.py
@@ -6,6 +6,8 @@ About
 
 Acquire PDF documents from USPTO servers.
 
+https://meta.ip-tools.org/t/uspto-introducing-ppubs-deprecating-patft-appft-pubeast-and-pubwest/178
+
 
 Synopsis
 ========
@@ -21,11 +23,8 @@ Synopsis
 Resources
 =========
 
-- US Published Patent Application Full-Page Images (ex. AppFT image server)
-  https://appft.uspto.gov/netahtml/PTO/appimg.htm
-
-- US Patent Full-Page Images (ex. PatFT image server)
-  https://patft.uspto.gov/netahtml/PTO/patimg.htm
+- USPTO Patent Public Search tool (PPUBS)
+  https://ppubs.uspto.gov/pubwebapp/
 
 
 Examples
@@ -35,45 +34,50 @@ Applications
 ------------
 
 - US20160105912
-  - :view: https://pdfaiw.uspto.gov/.aiw?docid=20160105912
+  - https://ppubs.uspto.gov/dirsearch-public/image-conversion/convert?url=us-pgpub/US/2016/0105/912/00000001.tif
 - US20160101909A1
-  - :full: https://pdfaiw.uspto.gov/fdd/09/2016/19/010/0.pdf
-  - :page: https://pdfaiw.uspto.gov/09/2016/19/010/2.pdf 
+  - https://ppubs.uspto.gov/dirsearch-public/image-conversion/convert?url=us-pgpub/US/2016/0101/909/00000001.tif
 - US2022110447
-  - :full: https://pdfaiw.uspto.gov/fdd/47/2022/04/011/0.pdf
-  - :page: https://pdfaiw.uspto.gov/47/2022/04/011/2.pdf
+  - https://ppubs.uspto.gov/dirsearch-public/image-conversion/convert?url=us-pgpub/US/2022/0110/447/00000001.tif
 
 Publications
 ------------
 
 - US9317610
-  - :view: https://pdfpiw.uspto.gov/.piw?docid=9317610
-- US10194689B2 
-  - :full: https://pdfpiw.uspto.gov/fdd/89/946/101/0.pdf
-  - :page: https://pdfpiw.uspto.gov/89/946/101/5.pdf
-- US05123456   
-  - :full: https://pdfpiw.uspto.gov/fdd/56/234/051/0.pdf
-  - :page: https://pdfpiw.uspto.gov/56/234/051/2.pdf
+  - https://ppubs.uspto.gov/dirsearch-public/image-conversion/convert?url=uspat/US/09/317/610/00000001.tif
+- US10194689B2
+  - https://ppubs.uspto.gov/dirsearch-public/image-conversion/convert?url=uspat/US/G89/946/101/00000001.tif
+- US05123456
+  - https://ppubs.uspto.gov/dirsearch-public/image-conversion/convert?url=uspat/US/05/123/456/00000001.tif
 
 """
 import logging
 import os
 import socket
 import sys
+import tempfile
 from collections import OrderedDict
+from datetime import timedelta
 
-import requests
-from bs4 import BeautifulSoup
+import requests_cache
 from beaker.cache import cache_region
 from pyramid.httpexceptions import HTTPNotFound
+from repoze.lru import lru_cache
 
 from patzilla.util.config import to_list
+from patzilla.util.image.convert import pdf_join
 from patzilla.util.network.browser import regular_user_agent
 from patzilla.util.numbers.normalize import normalize_patent
 
 logger = logging.getLogger(__name__)
 
-http = requests.Session()
+http = requests_cache.CachedSession(
+    "patzilla-uspto-cache",
+    cache_control=False,  # Use Cache-Control headers for expiration, if available
+    expire_after=timedelta(days=1),  # Otherwise expire responses after one day
+    match_headers=False,
+)
+http.headers["User-Agent"] = regular_user_agent
 
 
 # Timeout for HTTP requests.
@@ -83,9 +87,33 @@ DEFAULT_TIMEOUT = 5.0
 class UsptoPdfSection:
     FULL = -1
     FRONT_PAGE = 1
-    DRAWINGS = 2
-    SPECIFICATIONS = 3
-    CLAIMS = 4
+    BIBLIO = 2
+    ABSTRACT = 3
+    DRAWINGS = 4
+    SPECIFICATIONS = 5
+    DESCRIPTION = 6
+    CLAIMS = 7
+    SEARCH_REPORT = 8
+    SUPPLEMENTAL = 9
+    PTAB = 10
+    CERT_CORRECTION = 11
+    CERT_REEXAMINATION = 12
+
+
+UsptoPdfSectionAttributeMap = {
+    UsptoPdfSection.FRONT_PAGE: "frontPage",
+    UsptoPdfSection.BIBLIO: "bib",
+    UsptoPdfSection.ABSTRACT: "abstract",
+    UsptoPdfSection.DRAWINGS: "drawings",
+    UsptoPdfSection.SPECIFICATIONS: "specification",
+    UsptoPdfSection.DESCRIPTION: "description",
+    UsptoPdfSection.CLAIMS: "claims",
+    UsptoPdfSection.SEARCH_REPORT: "searchReport",
+    UsptoPdfSection.SUPPLEMENTAL: "supplemental",
+    UsptoPdfSection.PTAB: "ptab",
+    UsptoPdfSection.CERT_CORRECTION: "certCorrection",
+    UsptoPdfSection.CERT_REEXAMINATION: "certReexamination",
+}
 
 
 class UsptoPdfReferenceType:
@@ -109,41 +137,34 @@ def get_reference_type(document):
     return reference_type
 
 
-def document_viewer_url(document_number, validate=True):
+def document_viewer_url(document_number):
+    """
+    Compute URL to document number.
+    """
+    access = UsptoDocumentAccess(document_number=document_number)
+    url = access.get_page_url()
+    return {"location": url, "origin": "USPTO"}
 
-    document = normalize_patent(document_number, for_ops=False, as_dict=True, provider='uspto')
-    reference_type = get_reference_type(document)
-    url_tpl = None
-    if reference_type == UsptoPdfReferenceType.APPLICATION:
-        url_tpl = 'https://pdfaiw.uspto.gov/.aiw?docid={docid}'
 
-    elif reference_type == UsptoPdfReferenceType.PUBLICATION:
-        url_tpl = 'https://pdfpiw.uspto.gov/.piw?docid={docid}'
-
-    if url_tpl:
-        url = url_tpl.format(docid=document.number)
-
-        result = {'location': url, 'origin': 'USPTO'}
-
-        # Pre-flight check upstream url for existence of document.
-        # TODO: Read and propagate upstream error message like
-        #       `<b>Error Message: Unable to locate 0000000</b>`.
-        if validate:
-            try:
-                response = fetch_url(url)
-                print("response.content:", response.content)
-                if 'is not a valid ID' in response.content or 'Unable to locate' in response.content:
-                    raise HTTPNotFound('USPTO [PDF]: Resource at {} not found'.format(url))
-            except HTTPNotFound:
-                raise
-            except Exception as ex:  # pragma: nocover
-                raise ValueError('USPTO [PDF]: Failed to access resource at {}. Exception: {}'.format(url, ex))
-
-        return result
+@cache_region('static')
+def url_to_pdf(url):
+    """
+    Acquire page as PNG, and convert to PDF, cached.
+    """
+    tmpname = tempfile.mktemp(suffix=".pdf")
+    command = "convert '{}' '{}'".format(url, tmpname)
+    os.system(command)
+    with open(tmpname, mode="rb") as f:
+        payload = f.read()
+    os.unlink(tmpname)
+    return payload
 
 
 @cache_region('static')
 def fetch_url(url, method="get"):
+    """
+    Fetch response for URL.
+    """
 
     # Use default timeout.
     timeout = DEFAULT_TIMEOUT
@@ -172,103 +193,218 @@ def fetch_url(url, method="get"):
     raise HTTPNotFound(msg)
 
 
-def pdf_index(document_number, include=None):
+def get_patent_number_patentcenter(any_number):
+    """
+    Inquire patent number from `patentcenter.uspto.gov`.
 
-    if include:
-        include = to_list(include)
-
-    # Normalize document number and compute reference type (application vs. publication).
-    document = normalize_patent(document_number, for_ops=False, as_dict=True, provider='uspto')
+    TODO: This will quickly trigger `Service Unavailable  [PUBLIC], Max search limit reached.`.
+          So, find another way to resolve USPTO application or publication numbers to patent numbers.
+    """
+    document = normalize_patent(any_number, for_ops=False, as_dict=True, provider='uspto-ng')
+    if not document:
+        return
     reference_type = get_reference_type(document)
 
-    # Compute index page URL.
+    # US20140071615A1
     if reference_type == UsptoPdfReferenceType.APPLICATION:
-        baseurl = 'https://pdfaiw.uspto.gov'
-        pageurl = '/.aiw?docid={}'
+        url_prefix = "https://patentcenter.uspto.gov/retrieval/public/v2/application/data?publicationNumber="
+
+    # US07261559
     elif reference_type == UsptoPdfReferenceType.PUBLICATION:
-        baseurl = 'https://pdfpiw.uspto.gov'
-        pageurl = '/.piw?docid={}'
+        url_prefix = "https://patentcenter.uspto.gov/retrieval/public/v2/application/data?applicationNumberText="
 
-    url = baseurl + pageurl.format(document.number)
+    else:
+        raise TypeError("Unknown document type: {}".format(reference_type))
 
-    logger.info("Acquiring document index at {}".format(url))
+    start_url = url_prefix + document.number
+    logger.info("Inquiring information at: {}".format(start_url))
+    try:
+        response = http.get(start_url).json()
+        if "error" in response:
+            raise ValueError("Acquiring data failed. Reason: {}, {}".format(response["error"], response["message"]))
+        return response["applicationMetaData"]["patentNumber"]
+    except KeyError as ex:
+        raise ValueError("USPTO document number not found or incorrect format: {}".format(any_number))
 
-    # Fetch and parse HTML pages to create map of section => URLs.
-    # <embed src="//pdfpiw.uspto.gov/89/946/101/5.pdf" width="100%" height="850" type=application/pdf name="10194689" ></embed>
-    section_url_map = OrderedDict()
-    for section in [UsptoPdfSection.FRONT_PAGE, UsptoPdfSection.DRAWINGS, UsptoPdfSection.SPECIFICATIONS, UsptoPdfSection.CLAIMS]:
 
-        # When `include` option is given, only inquire specified sections to save resources.
-        if include and section not in include:
-            continue
+class UsptoDocumentAccess:
 
-        # Fetch single HTML page.
-        url_to_section = url + "&SectionNum={}".format(section)
-        response = fetch_url(url_to_section)
+    def __init__(self, document_number):
+        self.document_number = document_number
+        self.document = normalize_patent(self.document_number, for_ops=False, as_dict=True, provider="uspto-ng")
 
-        # Read `src` attribute from `<embed>` element.
-        soup = BeautifulSoup(response.content, features="lxml")
-        embed = soup.find("embed", attrs={"type": "application/pdf"})
-        embed_src = embed["src"]
+    @lru_cache(maxsize=None)
+    def inquire_images(self):
+        """
+        Inquire patent number from `ppubs.uspto.gov`.
+        """
+        if not self.document:
+            return
 
-        # Compute full URL to PDF document.
-        # If any host name (`uspto.gov`) is already present in the URL, it means the `embed.src`
-        # is something like `//pdfpiw.uspto.gov/59/615/072/3.pdf`. In this case, make sure to
-        # prefix it with the HTTP scheme `https:`.
-        if "uspto.gov" in embed_src:
-            pdfurl = "https:" + embed_src
+        # `07270540` would be an invalid query, so strip leading zeros.
+        logger.info("Inquiring document: {}".format(self.document.number))
+        query = '("{}").pn.'.format(self.document.number)
+
+        data = {
+            "start": 0,
+            "pageCount": 500,
+            "sort": "date_publ desc",
+            "docFamilyFiltering": "familyIdFiltering",
+            "searchType": 1,
+            "familyIdEnglishOnly": True,
+            "familyIdFirstPreferred": "US-PGPUB",
+            "familyIdSecondPreferred": "USPAT",
+            "familyIdThirdPreferred": "FPRS",
+            "showDocPerFamilyPref": "showEnglish",
+            "queryId": 0,
+            "tagDocSearch": False,
+            "query": {
+                # "caseId": 1623140,
+                "hl_snippets": "2",
+                "op": "OR",
+                "q": query,
+                "queryName": query,
+                "highlights": "1",
+                "qt": "brs",
+                "spellCheck": False,
+                "viewName": "tile",
+                "plurals": True,
+                "britishEquivalents": True,
+                "databaseFilters": [
+                    {"databaseName": "US-PGPUB", "countryCodes": []},
+                    {"databaseName": "USPAT", "countryCodes": []},
+                    {"databaseName": "USOCR", "countryCodes": []},
+                ],
+                "searchType": 1,
+                "ignorePersist": True,
+                "userEnteredQuery": query,
+            },
+        }
+
+        response = http.post("https://ppubs.uspto.gov/dirsearch-public/searches/searchWithBeFamily", json=data)
+        data = response.json()
+
+        # print(json.dumps(data, indent=2))
+
+        if "error" in data and data["error"]:
+            logger.error(data["error"])
+
+        return data
+
+    def get_page_url(self, page=1):
+        """
+        Get URL to PNG image, by page.
+        """
+
+        if page == 0:
+            return
+
+        data = self.inquire_images()
+        if data["patents"]:
+            first_hit = data["patents"][0]
+            return _get_image_url(first_hit, page=page)
         else:
-            pdfurl = baseurl + embed_src
-        section_url_map[section] = pdfurl
+            message = "USPTO document number not found or incorrect format: {}".format(self.document_number)
+            logger.warning(message)
 
-    if not section_url_map:
-        logger.warning("Unable to compute PDF section map for document {}".format(document_number))
+            msg = 'USPTO [PDF]: Resource {} not found'.format(self.document_number)
+            raise HTTPNotFound(msg)
 
-    return section_url_map
+    def get_allpages_urls(self):
+        """
+        Build list of URLs for all pages of a PDF document.
+        """
+        data = self.inquire_images()
+
+        if not (data and data["patents"]):
+            msg = 'USPTO [PDF]: Resource {} not found'.format(self.document_number)
+            raise HTTPNotFound(msg)
+
+        first_hit = data["patents"][0]
+        page_count = first_hit["pageCount"]
+        urls = []
+        for page in range(1, page_count + 1):
+            url = _get_image_url(first_hit, page=page)
+            urls.append(url)
+        return urls
+
+    def pdf_index(self, include=None):
+        """
+        Build map from document section to (start) URL.
+        """
+
+        if include:
+            include = to_list(include)
+
+        # Compute index page URL.
+        section_url_map = OrderedDict()
+        data = self.inquire_images()
+
+        if data["patents"]:
+            first_hit = data["patents"][0]
+
+            for section, attrprefix in UsptoPdfSectionAttributeMap.items():
+
+                # When `include` option is given, only inquire specified sections to save resources.
+                if include and section not in include:
+                    continue
+
+                start = first_hit.get(attrprefix + "Start")
+                end = first_hit.get(attrprefix + "End")
+
+                url = self.get_page_url(page=start)
+                section_url_map[section] = url
+
+        if not section_url_map:
+            logger.warning("Unable to compute PDF section map for document {}".format(self.document_number))
+
+        # logger.info("section_url_map: {}".format(section_url_map))
+
+        return section_url_map
 
 
-def pdf_url(document_number, section=UsptoPdfSection.FULL):
+def _get_image_url(patent_data, page=1):
+    address = "{}/{}.tif".format(patent_data["imageLocation"], str(page).zfill(8))
+    url = "https://ppubs.uspto.gov/dirsearch-public/image-conversion/convert?url={}".format(address)
+    logger.debug("Computed URL: {}".format(url))
+    return url
+
+
+def get_page_url_by_patent_number(pn, page=1):
+    address = "uspat/US/{}/{}/{}/{}.tif".format(pn[:-6].zfill(2), pn[-6:-3], pn[-3:], str(page).zfill(8))
+    url = 'https://ppubs.uspto.gov/dirsearch-public/image-conversion/convert?url={}'.format(address)
+    return url
+
+
+def png_url(document_number, section=UsptoPdfSection.DRAWINGS):
     """
     # Application
-    >>> pdf_url('US2016101909A1')
-    'https://pdfaiw.uspto.gov/fdd/09/2016/19/010/0.pdf'
+    >>> png_url('US2016101909A1', section=UsptoPdfSection.DRAWINGS)
+    'https://ppubs.uspto.gov/dirsearch-public/image-conversion/convert?url=us-pgpub/US/2016/0101/909/00000002.tif'
 
     # Grant I
-    >>> pdf_url('US10194689B2')
-    'https://pdfpiw.uspto.gov/fdd/89/946/101/0.pdf'
+    >>> png_url('US10194689B2', section=UsptoPdfSection.DESCRIPTION)
+    'https://ppubs.uspto.gov/dirsearch-public/image-conversion/convert?url=uspat/US/G89/946/101/00000015.tif'
 
     # Grant II
-    >>> pdf_url('US2548918')
-    'https://pdfpiw.uspto.gov/fdd/18/489/025/0.pdf'
+    >>> png_url('US2548918', section=UsptoPdfSection.CLAIMS)
+    'https://ppubs.uspto.gov/dirsearch-public/image-conversion/convert?url=uspat/US/02/548/918/00000003.tif'
     """
 
-    document = normalize_patent(document_number, for_ops=False, as_dict=True, provider='uspto')
+    logger.debug("Computing PDF URL for {}".format(document_number))
+
+    document = normalize_patent(document_number, for_ops=False, as_dict=True, provider="uspto-ng")
+    logger.debug("Normalized document number: {}".format(document))
     if not document:
         return
 
-    fdd_infix = ""
     if section == UsptoPdfSection.FULL:
-        fdd_infix = "fdd/"
-        page = 0
+        raise NotImplementedError("As of 2022, there is no direct URL to the full PDF document at USPTO")
     else:
-        index = pdf_index(document, include=UsptoPdfSection.DRAWINGS)
+        access = UsptoDocumentAccess(document_number=document_number)
+        index = access.pdf_index()
         url = index.get(section)
         return url
-
-    # Application
-    if len(document.number) == 11:
-        n = document.number
-        url = 'https://pdfaiw.uspto.gov/{}{}/{}/{}/{}/{}.pdf'.format(fdd_infix, n[9:11], n[0:4], n[7:9], n[4:7], page)
-
-    # Publication and Grant
-    elif len(document.number) == 8:
-        n = document.number
-        url = 'https://pdfpiw.uspto.gov/{}{}/{}/{}/{}.pdf'.format(fdd_infix, n[6:8], n[3:6], n[0:3], page)
-
-    else:  # pragma: nocover
-        raise ValueError('US document number "{}" has unexpected format'.format(document_number))
-
-    return url
 
 
 def fetch_pdf(document_number, section=UsptoPdfSection.FULL):
@@ -278,37 +414,58 @@ def fetch_pdf(document_number, section=UsptoPdfSection.FULL):
     Blueprint addresses:
 
     - Application: US2016101909A1
-      :full: https://pdfaiw.uspto.gov/fdd/09/2016/19/010/0.pdf
-      :page: https://pdfaiw.uspto.gov/09/2016/19/010/2.pdf
+      https://ppubs.uspto.gov/dirsearch-public/image-conversion/convert?url=us-pgpub/US/2016/0101/909/00000001.tif
 
     - Grant: US10194689B2
-      :full: https://pdfpiw.uspto.gov/fdd/89/946/101/0.pdf
-      :page: https://pdfpiw.uspto.gov/89/946/101/5.pdf
+      https://ppubs.uspto.gov/dirsearch-public/image-conversion/convert?url=uspat/US/G89/946/101/00000001.tif
     """
+    logger.info("Fetching PDF for {}".format(document_number))
+    access = UsptoDocumentAccess(document_number=document_number)
 
-    url = pdf_url(document_number, section=section)
+    if section == UsptoPdfSection.FULL:
+        pdf_payloads = []
+        urls = access.get_allpages_urls()
+        for url in urls:
+            logger.info("Fetching page from {}".format(url))
+            pdf = url_to_pdf(url)
+            pdf_payloads.append(pdf)
+        outcome = pdf_join(pdf_payloads)
+        return outcome
 
-    logger.info('Accessing document {} at {}'.format(document_number, url))
-    response = fetch_url(url)
+    else:
 
-    # Debugging
-    # print('response.status:  {}'.format(response.status_code))
-    # print('response.content: {}'.format(response.content))
+        raise NotImplementedError("Conversion of single pages from PNG to PDF not implemented yet")
 
-    return response.content
+
+def fetch_png(document_number, section=UsptoPdfSection.FULL):
+    if section == UsptoPdfSection.FULL:
+        raise NotImplementedError("Acquiring a full document as PNG not possible")
+    else:
+        url = png_url(document_number, section=section)
+        logger.info('Accessing document {} at {}'.format(document_number, url))
+        response = fetch_url(url)
+
+        # Debugging
+        # print('response.status:  {}'.format(response.status_code))
+        # print('response.content: {}'.format(response.content))
+
+        return response.content
 
 
 example_numbers = [
     # Applications
+    # https://patentcenter.uspto.gov/retrieval/public/v2/application/data?publicationNumber=US20160101909A1
     'US20140071615A1',
     'US20140075562A1',
     'US2016101909A1',
     'US20070231208A1', 'US2007231208A1',
 
     # Publications
+    # https://patentcenter.uspto.gov/retrieval/public/v2/application/data?applicationNumberText=07261559
     # 7-char numbers
+    'US5051145A',
     'US7261559B2',
-    'US07270540B2',
+    'US07270540B2',     # Access `childContinuityBag[0][patentNumber]==5101689`?
     'US7493690B2',
     'US7494593B1', 'US7494597B2',
     # 8-char numbers
@@ -331,6 +488,12 @@ def run_examples(fun, ev=None):  # pragma: nocover
 if __name__ == '__main__':  # pragma: nocover
     """
     Demo program for computing the direct access URL to PDF document(s) on USPTO servers.
+
+    Synopsis::
+
+        python -m patzilla.access.uspto.pdf
+        python -m patzilla.access.uspto.pdf US20140071638A1
+        python -m patzilla.access.uspto.pdf US20070231208A1
     """
 
     from patzilla.boot.cache import configure_cache_backend
@@ -346,4 +509,7 @@ if __name__ == '__main__':  # pragma: nocover
         print(fetch_pdf(number))
 
     else:
-        run_examples(pdf_url)
+        def worker(document_number):
+            access = UsptoDocumentAccess(document_number=document_number)
+            return access.get_page_url()
+        run_examples(worker)