+++ CHANGES.rst
@@ -50,6 +50,7 @@ Development
 - [mw] Add CLI interface for basic operations on EPO/OPS
 - [mw] Fix access to DPMAregister due to upstream changes
 - [mw] Add CLI interface for search operation on IFI CLAIMS
+- [mw] Fix access to USPTO documents using USPTO PPUBS, PatFT and AppFT are deprecated
 
 
 2019-11-01 0.169.3